reorganized components, updated ui
