--- modelmux/src/components/chat-interface.tsx
@@ -1,175 +1,0 @@
-import { useState } from "react";
-import { Send, Plus, RotateCcw, MoreVertical } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import Input from "@/components/ui/input";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-  DropdownMenuSeparator,
-} from "@/components/ui/dropdown-menu";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { Response } from "@/lib/ollama";
-
-interface Message {
-  id: string;
-  content: string;
-  isUser: boolean;
-  timestamp: Date;
-}
-
-interface MemoryMessage {
-  role: "user" | "system" | "assistant";
-  content: string;
-}
-
-export function ChatInterface() {
-  const [messages, setMessages] = useState<Message[]>([
-    {
-      id: "1",
-      content: "Hello! I'm your AI assistant. How can I help you today?",
-      isUser: false,
-      timestamp: new Date(),
-    },
-  ]);
-
-  const [inputValue, setInputValue] = useState("");
-  const [memory, setMemory] = useState<MemoryMessage[]>([]);
-
-  const handleSendMessage = () => {
-    if (!inputValue.trim()) return;
-
-    const userMessage: Message = {
-      id: Date.now().toString(),
-      content: inputValue,
-      isUser: true,
-      timestamp: new Date(),
-    };
-
-    setMemory((prev) => [...prev, { role: "user", content: inputValue }]);
-    setMessages((prev) => [...prev, userMessage]);
-
-    // LLM response
-
-    async function run() {
-      const { message, data } = await Response(
-        { role: "user", content: inputValue },
-        memory
-      );
-      setMemory((prev) => [...prev, { role: "assistant", content: message }]);
-      setMessages((prev) => [...prev, data]);
-    }
-    run();
-    setInputValue("");
-  };
-
-  const handleKeyPress = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter" && !e.shiftKey) {
-      e.preventDefault();
-      handleSendMessage();
-    }
-  };
-
-  return (
-    <div className="flex flex-col h-screen bg-background">
-      {/* Header */}
-      <div className="flex items-center justify-between p-4 border-b">
-        <h1 className="text-xl font-semibold">ModelMux Chat</h1>
-        <Button variant="outline" size="sm">
-          <Plus className="h-4 w-4 mr-2" />
-          New Chat
-        </Button>
-      </div>
-
-      {/* Messages Area */}
-      <ScrollArea className="flex-1 p-4">
-        <div className="space-y-4 max-w-4xl mx-auto">
-          {messages.map((message) => (
-            <div
-              key={message.id}
-              className={`flex ${
-                message.isUser ? "justify-end" : "justify-start"
-              }`}
-            >
-              <div
-                className={`max-w-[80%] rounded-lg p-4 relative group ${
-                  message.isUser
-                    ? "bg-primary text-primary-foreground"
-                    : "bg-muted"
-                }`}
-              >
-                <p className="text-sm leading-relaxed">{message.content}</p>
-                <span className="text-xs opacity-70 mt-2 block">
-                  {message.timestamp.toLocaleTimeString()}
-                </span>
-                {/* dropdown menu for retry icon*/}
-                {!message.isUser && (
-                  <div className="absolute bottom-2 right-2 opacity-0 group-hover:opacity-100 transition-opacity">
-                    <DropdownMenu>
-                      <Tooltip>
-                        <DropdownMenuTrigger asChild>
-                          <TooltipTrigger asChild>
-                            <button
-                              className="p-1 rounded hover:bg-accent focus:outline-none"
-                              title="More"
-                            >
-                              <RotateCcw className="h-4 w-4" />
-                            </button>
-                          </TooltipTrigger>
-                        </DropdownMenuTrigger>
-                        <TooltipContent side="bottom" align="center">
-                          Retry message
-                        </TooltipContent>
-                      </Tooltip>
-                      <DropdownMenuContent align="end">
-                        <DropdownMenuItem
-                          onSelect={(e) => e.preventDefault()}
-                          className="flex items-center gap-2"
-                        >
-                          <RotateCcw className="h-3 w-3" />
-                          Retry
-                        </DropdownMenuItem>
-                        <DropdownMenuSeparator />
-                        <DropdownMenuItem onSelect={(e) => e.preventDefault()}>
-                          Copy message
-                        </DropdownMenuItem>
-                        <DropdownMenuItem onSelect={(e) => e.preventDefault()}>
-                          Report issue
-                        </DropdownMenuItem>
-                      </DropdownMenuContent>
-                    </DropdownMenu>
-                  </div>
-                )}
-              </div>
-            </div>
-          ))}
-        </div>
-      </ScrollArea>
-
-      {/* Input Area */}
-      <div className="p-4 border-t bg-background sticky bottom-0">
-        <div className="max-w-4xl mx-auto">
-          <div className="flex gap-2">
-            <Input
-              value={inputValue}
-              type={"text"}
-              onChange={(e) => setInputValue(e.target.value)}
-              onKeyDown={handleKeyPress}
-              placeholder="Type your message here..."
-              className="flex-1"
-            />
-            <Button onClick={handleSendMessage} disabled={!inputValue.trim()}>
-              <Send className="h-4 w-4" />
-            </Button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
